Refactor codes for using common codes Output to `bin` folder as the following |. |   global-inject-demo.exe |   global-inject-demo_32.exe |   global-inject-demo_D.exe |   global-inject-demo_D_32.exe | +---32 |       global-inject-lib.dll |       global-inject-lib_D.dll | +---64         global-inject-lib.dll         global-inject-lib_D.dll
