--- global-inject-demo/functions.h
@@ -1,4 +1,0 @@
-#pragma once
-
-BOOL SetPrivilege(HANDLE hToken, LPCTSTR lpszPrivilege, BOOL bEnablePrivilege);
-BOOL SetDebugPrivilege(BOOL bEnablePrivilege);
