+++ global-inject-lib/dll_inject.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "dll_inject.h"
 #include "logger.h"
+#include "functions.h"
 
 extern HINSTANCE g_hDllInst;
 
@@ -136,227 +137,6 @@ namespace
 		"\x41\x5C\x5F\x5E\x5B\x5D\xC3";
 
 	constexpr SIZE_T x64ShellcodeSize = sizeof(x64Shellcode) - 1;
-
-	//
-	// https://docs.microsoft.com/en-us/windows/win32/sysinfo/verifying-the-system-version
-	//
-	BOOL CheckWindowsVersion(DWORD dwMajorVersion, DWORD dwMinorVersion,
-		WORD wServicePackMajor, WORD wServicePackMinor, int op)
-	{
-		// Initialize the OSVERSIONINFOEX structure
-		OSVERSIONINFOEX osvi;
-
-		ZeroMemory(&osvi, sizeof(OSVERSIONINFOEX));
-		osvi.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
-		osvi.dwMajorVersion = dwMajorVersion;
-		osvi.dwMinorVersion = dwMinorVersion;
-		osvi.wServicePackMajor = wServicePackMajor;
-		osvi.wServicePackMinor = wServicePackMinor;
-
-		// Initialize the type mask
-		DWORD dwTypeMask = VER_MAJORVERSION | VER_MINORVERSION |
-			VER_SERVICEPACKMAJOR | VER_SERVICEPACKMINOR;
-
-		// Initialize the condition mask
-		DWORDLONG dwlConditionMask = 0;
-
-		VER_SET_CONDITION(dwlConditionMask, VER_MAJORVERSION, op);
-		VER_SET_CONDITION(dwlConditionMask, VER_MINORVERSION, op);
-		VER_SET_CONDITION(dwlConditionMask, VER_SERVICEPACKMAJOR, op);
-		VER_SET_CONDITION(dwlConditionMask, VER_SERVICEPACKMINOR, op);
-
-		// Perform the test
-		return VerifyVersionInfo(&osvi, dwTypeMask, dwlConditionMask);
-	}
-
-	//
-	// Based on:
-	// http://securityxploded.com/ntcreatethreadex.php
-	//
-	HANDLE MyCreateRemoteThread(HANDLE hProcess,
-		LPTHREAD_START_ROUTINE lpStartAddress, LPVOID lpParameter, USHORT targetProcessArch)
-	{
-#ifndef _WIN64
-		if (targetProcessArch == IMAGE_FILE_MACHINE_AMD64) {
-			// WOW64 to x64 native, use heaven's gate.
-			return (HANDLE)CreateRemoteThread64(
-				HANDLE_TO_DWORD64(hProcess), PTR_TO_DWORD64(lpStartAddress), PTR_TO_DWORD64(lpParameter));
-		}
-#endif // _WIN64
-
-		using NtCreateThreadEx_t = NTSTATUS(WINAPI*)(
-			OUT PHANDLE hThread,
-			IN ACCESS_MASK DesiredAccess,
-			IN LPVOID ObjectAttributes,
-			IN HANDLE ProcessHandle,
-			IN LPTHREAD_START_ROUTINE lpStartAddress,
-			IN LPVOID lpParameter,
-			IN BOOL CreateSuspended,
-			IN ULONG_PTR StackZeroBits,
-			IN ULONG_PTR SizeOfStackCommit,
-			IN ULONG_PTR SizeOfStackReserve,
-			OUT LPVOID lpBytesBuffer
-			);
-
-		static NtCreateThreadEx_t pNtCreateThreadEx = []() {
-			// Only needed for Windows Vista and 7.
-			if (CheckWindowsVersion(6, 0, 0, 0, VER_GREATER_EQUAL) &&
-				!CheckWindowsVersion(6, 2, 0, 0, VER_GREATER_EQUAL)) {
-				HMODULE hNtdll = GetModuleHandle(L"ntdll.dll");
-				if (hNtdll) {
-					return (NtCreateThreadEx_t)GetProcAddress(hNtdll, "NtCreateThreadEx");
-				}
-			}
-
-			return (NtCreateThreadEx_t)nullptr;
-		}();
-
-		LPSECURITY_ATTRIBUTES lpThreadAttributes = nullptr;
-		DWORD dwCreationFlags = 0;
-
-		if (pNtCreateThreadEx) {
-			HANDLE hThread;
-			ULONG_PTR bOutBuffer1[2];
-			ULONG_PTR bOutBuffer2[1];
-			struct {
-				ULONG_PTR Size;
-				ULONG_PTR Unknown1;
-				ULONG_PTR nBuf1Size;
-				void* pBuf1;
-				ULONG_PTR Unknown2;
-				ULONG_PTR Unknown3;
-				ULONG_PTR nBuf2Size;
-				void* pBuf2;
-				ULONG_PTR Unknown4;
-			} param = { sizeof(param), 0x10003, sizeof(ULONG_PTR) * 2, bOutBuffer1, 0, 0x10004, sizeof(ULONG_PTR), bOutBuffer2, 0 };
-
-			NTSTATUS result = pNtCreateThreadEx(&hThread, 0x1FFFFF, lpThreadAttributes,
-				hProcess, lpStartAddress, lpParameter, (dwCreationFlags & CREATE_SUSPENDED) ? TRUE : FALSE, 0, 0, 0, &param);
-			if (result < 0) {
-				SetLastError(LsaNtStatusToWinError(result));
-				return nullptr;
-			}
-
-			return hThread;
-		}
-
-		return CreateRemoteThread(hProcess, lpThreadAttributes, 0, lpStartAddress, lpParameter, dwCreationFlags, nullptr);
-	}
-
-	//
-	// Reference: https://repnz.github.io/posts/apc/wow64-user-apc/
-	//
-	ULONG64 EncodeWow64ApcRoutine(ULONG64 ApcRoutine)
-	{
-		return (ULONG64)((-(INT64)ApcRoutine) << 2);
-	}
-
-	//
-	// Reference: https://repnz.github.io/posts/apc/user-apc/
-	//
-	BOOL MyQueueUserAPC(PAPCFUNC pfnAPC, HANDLE hThread, ULONG_PTR dwData, USHORT targetProcessArch)
-	{
-#ifndef _WIN64
-		if (targetProcessArch == IMAGE_FILE_MACHINE_AMD64) {
-			// WOW64 to x64 native, use heaven's gate.
-			//
-			// "Microsoft added a validation to prevent a programming error:
-			// If you try to queue an APC from a 32 bit process to a 64 bit
-			// process and you use a 32 bit address, you'll get this status code:
-			// [...] STATUS_INVALID_HANDLE"
-			// https://repnz.github.io/posts/apc/wow64-user-apc/
-			return NtQueueApcThread64(
-				HANDLE_TO_DWORD64(hThread), PTR_TO_DWORD64(pfnAPC), (DWORD64)dwData, 0, 0);
-		}
-#endif // _WIN64
-
-		using NtQueueApcThread_t = DWORD(WINAPI*)(
-			IN HANDLE ThreadHandle,
-			IN PVOID ApcDispatchRoutine,
-			IN ULONG_PTR SystemArgument1,
-			IN ULONG_PTR SystemArgument2,
-			IN ULONG_PTR SystemArgument3
-			);
-
-		static NtQueueApcThread_t pNtQueueApcThread = []() {
-			HMODULE hNtdll = GetModuleHandle(L"ntdll.dll");
-			if (hNtdll) {
-				return (NtQueueApcThread_t)GetProcAddress(hNtdll, "NtQueueApcThread");
-			}
-
-			return (NtQueueApcThread_t)nullptr;
-		}();
-
-		if (!pNtQueueApcThread) {
-			SetLastError(ERROR_PROC_NOT_FOUND);
-			return FALSE;
-		}
-
-#ifdef _WIN64
-		if (targetProcessArch == IMAGE_FILE_MACHINE_I386) {
-			// x64 native to WOW64, encode address.
-			pfnAPC = (PAPCFUNC)EncodeWow64ApcRoutine((ULONG64)pfnAPC);
-		}
-#endif // _WIN64
-
-		NTSTATUS result = pNtQueueApcThread(hThread, pfnAPC, dwData, 0, 0);
-		if (result < 0) {
-			SetLastError(LsaNtStatusToWinError(result));
-			return FALSE;
-		}
-
-		return TRUE;
-	}
-
-	USHORT GetProcessArch(HANDLE hProcess)
-	{
-		// For now, only IMAGE_FILE_MACHINE_I386 and IMAGE_FILE_MACHINE_AMD64.
-		// TODO: Use IsWow64Process2 if available.
-
-#ifndef _WIN64
-		SYSTEM_INFO siSystemInfo;
-		GetNativeSystemInfo(&siSystemInfo);
-		if (siSystemInfo.wProcessorArchitecture == PROCESSOR_ARCHITECTURE_INTEL) {
-			// 32-bit machine, only one option.
-			return IMAGE_FILE_MACHINE_I386;
-		}
-#endif // _WIN64
-
-		BOOL bIsWow64Process;
-		if (IsWow64Process(hProcess, &bIsWow64Process) && bIsWow64Process) {
-			return IMAGE_FILE_MACHINE_I386;
-		}
-
-		return IMAGE_FILE_MACHINE_AMD64;
-	}
-
-	std::filesystem::path GetEnginePath(USHORT machine)
-	{
-		std::filesystem::path libraryPath = wil::GetModuleFileName<std::wstring>(g_hDllInst);
-
-		auto folderPath = libraryPath.parent_path();
-		auto folderName = folderPath.filename();
-
-		if (folderName != L"32" && folderName != L"64") {
-			throw std::runtime_error("DLL file not in \\32 or \\64 folder");
-		}
-
-		PCWSTR newFolderName;
-		switch (machine) {
-		case IMAGE_FILE_MACHINE_I386:
-			newFolderName = L"32";
-			break;
-
-		case IMAGE_FILE_MACHINE_AMD64:
-			newFolderName = L"64";
-			break;
-
-		default:
-			throw std::logic_error("Unknown architecture");
-		}
-
-		return folderPath.parent_path() / newFolderName;
-	}
 }
 
 namespace DllInject
@@ -389,7 +169,7 @@ namespace DllInject
 			throw std::logic_error("Invalid architecture value");
 		}
 
-		std::wstring dllPath = GetEnginePath(targetProcessArch) / L"global-inject-lib.dll";
+		std::wstring dllPath = GetEnginePath(&targetProcessArch) / GetDllFileName();
 		size_t dllPathBytes = (dllPath.length() + 1) * sizeof(WCHAR);
 
 		HANDLE hRemoteSessionManagerProcess;
