--- global-inject-lib/functions.h
@@ -1,5 +1,0 @@
-#pragma once
-
-BOOL GetFullAccessSecurityDescriptor(
-    _Outptr_ PSECURITY_DESCRIPTOR* SecurityDescriptor,
-    _Out_opt_ PULONG SecurityDescriptorSize);
